+++ src/main.rs
@@ -1,13 +1,13 @@
 #![forbid(unsafe_code)]
 #![forbid(warnings)]
 
-use failure::Error;
+use anyhow::Result;
 use pretty_env_logger;
 use structopt::StructOpt;
 
 use netlify_ddns::{run, Args};
 
-fn main() -> Result<(), Error> {
+fn main() -> Result<()> {
     pretty_env_logger::init_timed();
     let args = Args::from_args();